+++ src/AtomUI.Desktop.Controls/QRCode/QRCode.SemanticParts.cs
@@ -1,0 +1,13 @@
+using AtomUI.Theme;
+using Avalonia.Controls;
+
+namespace AtomUI.Desktop.Controls;
+
+[SemanticPart(
+    "cover",
+    SelectorClass = "semantic-cover",
+    ContractType = typeof(Border),
+    Since = "6.0")]
+public partial class QRCode
+{
+}