+++ src/AtomUI.Desktop.Controls/QRCode/QRCode.cs
@@ -2,7 +2,7 @@ using AtomUI.Controls.Commons;
 
 namespace AtomUI.Desktop.Controls;
 
-public class QRCode : AbstractQRCode
+public partial class QRCode : AbstractQRCode
 {
     public QRCode()
     {